cambiar funcion de inicio de juego a otro archivo
